[AIE][AIEX] dma_channel_reset_for: self-contained re-arm binding + one lowering pass
Address review on #3393:

- objectfifo_rearm_binding carries the head bd_id + repeat per channel as
  optional head_bd_ids / repeat_counts. aie-assign-bd-ids folds them on right
  after it assigns the ids; the reset_for lowering reads them off the binding
  and no longer rescans the emitted aie.mem chain, so the op works from a
  hand-authored binding with no objectFIFO code behind it. Reject a shim tile
  in channel_tiles (host-managed, never re-armed here).

- Fold the dma_channel_reset_for expansion into aie-lower-dma-channel-reset:
  one pass expands reset_for then lowers every dma_channel_reset, run ahead of
  aie-dma-to-npu / aie-lower-set-lock since it emits push_queue / set_lock.
  Removes aie-lower-dma-channel-reset-for.
