+++ include/aie/Dialect/AIEX/Transforms/AIEXPasses.td
@@ -356,7 +356,22 @@ def AIELowerSetLock : Pass<"aie-lower-set-lock", "AIE::DeviceOp"> {
 }
 
 def AIELowerDmaChannelReset : Pass<"aie-lower-dma-channel-reset", "AIE::DeviceOp"> {
-  let summary = "Lowers all aiex.dma_channel_reset operations to npu.maskwrite32";
+  let summary = "Lowers aiex.dma_channel_reset(_for) to npu.maskwrite32 + re-arm";
+  let description = [{
+    Two concerns, one pass (kept together like the other reset lowerings):
+
+    - Each `aiex.dma_channel_reset_for @fifo` is first expanded, via the fifo's
+      `aie.objectfifo_rearm_binding`, into the resident re-arm of its non-shim
+      channels: an `aiex.dma_channel_reset` per channel, an `aiex.set_lock` per
+      producer/consumer lock, and an `aiex.npu.push_queue` START_QUEUE re-push
+      per channel (head BD id + repeat read straight off the binding, populated
+      by `--aie-assign-bd-ids`).
+    - Every `aiex.dma_channel_reset` -- standalone or just emitted above -- is
+      then lowered to the `aiex.npu.maskwrite32` reset pulse.
+
+    The emitted `set_lock` / `push_queue` are lowered later by
+    `--aie-lower-set-lock` / `--aie-dma-to-npu`, so run this before them.
+  }];
   let constructor = "xilinx::AIEX::createAIELowerDmaChannelResetPass()";
   let dependentDialects = [
     "mlir::arith::ArithDialect",
@@ -368,25 +383,6 @@ def AIELowerDmaChannelReset : Pass<"aie-lower-dma-channel-reset", "AIE::DeviceOp
 def AIELowerCoreReset : Pass<"aie-lower-core-reset", "AIE::DeviceOp"> {
   let summary = "Lowers all aiex.core_reset operations to npu.maskwrite32";
   let constructor = "xilinx::AIEX::createAIELowerCoreResetPass()";
-  let dependentDialects = [
-    "mlir::arith::ArithDialect",
-    "xilinx::AIE::AIEDialect",
-    "xilinx::AIEX::AIEXDialect",
-  ];
-}
-
-def AIELowerDmaChannelResetFor : Pass<"aie-lower-dma-channel-reset-for", "AIE::DeviceOp"> {
-  let summary = "Lowers aiex.dma_channel_reset_for to its re-arm trio";
-  let description = [{
-    Expands each `aiex.dma_channel_reset_for @fifo` into the resident re-arm of
-    the fifo's non-shim channels, resolved through the fifo's
-    `aie.objectfifo_rearm_binding`: an `aiex.dma_channel_reset` per channel, an
-    `aiex.set_lock` per producer/consumer lock, and an `aiex.npu.write32`
-    START_QUEUE re-push per channel (bd_id + repeat read from the resident
-    `aie.dma_start`). Run after `--aie-assign-bd-ids` and before
-    `--aie-lower-set-lock` / `--aie-lower-dma-channel-reset`.
-  }];
-  let constructor = "xilinx::AIEX::createAIELowerDmaChannelResetForPass()";
   let dependentDialects = [
     "mlir::arith::ArithDialect",
     "xilinx::AIE::AIEDialect",
